+++ src/component/mouse-dragdrop/mouse-dragdrop-cursor.component.ts
@@ -10,9 +10,9 @@
  */
 
 import { Component, ElementRef, ViewChild, ViewEncapsulation } from '@angular/core';
-import { DejaMouseDragDropService, IDragCursorInfos, IDropCursorInfos } from './mouse-dragdrop.service';
-import { Observable, BehaviorSubject } from 'rxjs/Rx';
+import { BehaviorSubject, Observable } from 'rxjs/Rx';
 import { Position } from './../../common/core/graphics/position';
+import { DejaMouseDragDropService, IDragCursorInfos, IDropCursorInfos } from './mouse-dragdrop.service';
 
 @Component({
     encapsulation: ViewEncapsulation.None,